+++ hibernate-core/src/test/java/org/hibernate/test/mapping/hhh14276/NestedIdClassDerivedIdentifiersTest.java
@@ -1,14 +1,15 @@
-package org.hibernate.test.mapping.hhh99999;
+package org.hibernate.test.mapping.hhh14276;
 
 import static org.hibernate.testing.transaction.TransactionUtil.doInJPA;
 
 import java.util.Map;
 
+import org.hibernate.Version;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.hql.spi.id.inline.InlineIdsOrClauseBulkIdStrategy;
 import org.hibernate.jpa.test.BaseEntityManagerFunctionalTestCase;
-import org.hibernate.test.mapping.hhh99999.entity.PlayerStat;
-import org.hibernate.test.mapping.hhh99999.entity.Score;
+import org.hibernate.test.mapping.hhh14276.entity.PlayerStat;
+import org.hibernate.test.mapping.hhh14276.entity.Score;
 import org.hibernate.testing.TestForIssue;
 import org.junit.Before;
 import org.junit.Test;
@@ -39,6 +40,7 @@ public class NestedIdClassDerivedIdentifiersTest extends BaseEntityManagerFuncti
         doInJPA( this::entityManagerFactory, em ->
         {
             // do nothing
+            System.out.println( "Hibernate version: " + Version.getVersionString() );
         } );
     }
