+++ hibernate-core/src/test/java/org/hibernate/test/mapping/hhh14276/entity/Score.java
@@ -1,4 +1,4 @@
-package org.hibernate.test.mapping.hhh99999.entity;
+package org.hibernate.test.mapping.hhh14276.entity;
 
 import java.io.Serializable;
 
@@ -74,48 +74,5 @@ public class Score implements Serializable
     public void setFinalScore(Integer finalScore)
     {
         this.finalScore = finalScore;
-    }
-
-    @Override
-    public int hashCode()
-    {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ( (gameId == null) ? 0 : gameId.hashCode() );
-        result = prime * result + ( (home == null) ? 0 : home.hashCode() );
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj)
-    {
-        if ( this == obj )
-            return true;
-        if ( obj == null )
-            return false;
-        if ( getClass() != obj.getClass() )
-            return false;
-        Score other = ( Score ) obj;
-        if ( gameId == null )
-        {
-            if ( other.gameId != null )
-                return false;
-        }
-        else if ( !gameId.equals( other.gameId ) )
-            return false;
-        if ( home == null )
-        {
-            if ( other.home != null )
-                return false;
-        }
-        else if ( !home.equals( other.home ) )
-            return false;
-        return true;
-    }
-
-    @Override
-    public String toString()
-    {
-        return "[" + gameId + ", " + home + ", " + rosterId + ", " + finalScore + "]";
     }
 }
