--- hibernate-core/src/test/java/org/hibernate/test/mapping/hhh99999/entity/ScoreId.java
@@ -1,79 +1,0 @@
-package org.hibernate.test.mapping.hhh99999.entity;
-
-import java.io.Serializable;
-
-public class ScoreId implements Serializable
-{
-    private static final long serialVersionUID = 1L;
-
-    private Integer gameId;
-
-    private Boolean home;
-
-    public ScoreId()
-    {
-    }
-
-    public Integer getGameId()
-    {
-        return gameId;
-    }
-
-    public void setGameId(Integer gameId)
-    {
-        this.gameId = gameId;
-    }
-
-    public Boolean getHome()
-    {
-        return home;
-    }
-
-    public void setHome(Boolean home)
-    {
-        this.home = home;
-    }
-
-    @Override
-    public int hashCode()
-    {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ( (gameId == null) ? 0 : gameId.hashCode() );
-        result = prime * result + ( (home == null) ? 0 : home.hashCode() );
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj)
-    {
-        if ( this == obj )
-            return true;
-        if ( obj == null )
-            return false;
-        if ( getClass() != obj.getClass() )
-            return false;
-        ScoreId other = ( ScoreId ) obj;
-        if ( gameId == null )
-        {
-            if ( other.gameId != null )
-                return false;
-        }
-        else if ( !gameId.equals( other.gameId ) )
-            return false;
-        if ( home == null )
-        {
-            if ( other.home != null )
-                return false;
-        }
-        else if ( !home.equals( other.home ) )
-            return false;
-        return true;
-    }
-
-    @Override
-    public String toString()
-    {
-        return "[" + gameId + ", " + home + "]";
-    }
-}
